Votyra: changed 2f images to 2i

+++ Assets/Votyra/Core/Images/Constraints/IImageConstraint2i.cs
@@ -4,6 +4,6 @@ namespace Votyra.Core.Images.Constraints
 {
     public interface IImageConstraint2i
     {
-        Range2i FixImage(Matrix2<float> editableMatrix, Range2i invalidatedImageArea, Direction direction);
+        Range2i FixImage(Matrix2<int> editableMatrix, Range2i invalidatedImageArea, Direction direction);
     }
 }
